Final commit after leatest updates validation was added to all screens action bar was removed and background color changed checking campaign number before logging in was added changing report screen and adapting it to campaign ressults

+++ app/src/main/java/com/example/ahmed/pointoflife/OnGetDataListener.java
@@ -3,7 +3,7 @@ package com.example.ahmed.pointoflife;
 import com.google.firebase.database.DataSnapshot;
 
 public interface OnGetDataListener {
-    void onSuccess(DataSnapshot dataSnapshot, double value, String counter);
+    void onSuccess(DataSnapshot dataSnapshot, int value, String counter);
     void onStart();
     void onFailure();
 }
